Fix: maxbuffer error and unhandled promises.
When certain errors would occur a promise rejection would trigger. These
were not getting handled and in the future this will kill the application.

When getting search results for a lot of entries we would fill up the
stdout buffer. This would lead to no output making it to the user. Going
to increase the max buffer size to see if this fixes it.